Added timing
The worker runs at a fixed framerate (default 30, can be changed at
runtime).
It will sleep for the remaining time.
If we are running late (keeps a running total) it will not sleep and
instead enable the fast_run flag in the envrionment.
This flag can be used to skip computation/loading if needed.